+++ packages/block-library/src/gallery/transforms.js
@@ -223,7 +223,7 @@ const transforms = {
 				if ( isGalleryV2Enabled() ) {
 					const innerBlocks = files.map( ( file ) =>
 						createBlock( 'core/image', {
-							url: createBlobURL( file ),
+							blob: createBlobURL( file ),
 						} )
 					);