+++ packages/editor/src/dataviews/types.ts
@@ -9,13 +9,27 @@ type PostStatus =
 
 export interface BasePost {
 	status?: PostStatus;
-	title: string | { rendered: string };
+	title: string | { rendered: string } | { raw: string };
 	type: string;
+	id: string | number;
 }
 export interface TemplateOrTemplatePart extends BasePost {
-	type: 'template' | 'template-part';
+	type: 'wp_template' | 'wp_template_part';
 	source: string;
 	has_theme_file: boolean;
+	id: string;
 }
 
-export type Post = TemplateOrTemplatePart | BasePost;
+export interface Pattern extends BasePost {
+	slug: string;
+	title: { raw: string };
+	content: {
+		raw: string;
+	};
+	wp_pattern_sync_status: string;
+}
+
+export type Post = TemplateOrTemplatePart | Pattern | BasePost;
+
+// Will be unnecessary after typescript 5.0 upgrade.
+export type CoreDataError = { message?: string; code?: string };