+++ storybook/stories/playground/index.story.js
@@ -16,10 +16,22 @@ export const _default = () => {
 	return <EditorFullPage />;
 };
 
+_default.parameters = {
+	sourceLink: 'storybook/stories/playground/fullpage/index.js',
+};
+
 export const Box = () => {
 	return <EditorBox />;
 };
 
+Box.parameters = {
+	sourceLink: 'storybook/stories/playground/box/index.js',
+};
+
 export const UndoRedo = () => {
 	return <EditorWithUndoRedo />;
+};
+
+UndoRedo.parameters = {
+	sourceLink: 'storybook/stories/playground/with-undo-redo/index.js',
 };